Don't use the local lint configuration with Hound
The Hound FAQ recommends converting a codebase to conform to the style guide over time, one pull request at a time. While linter exceptions must be made for the local, full-source lint, it is better to lint each PR without the local exceptions, to encourage each PR to improve the style of the code that it touches.