Remove the cross-module Generator initializer, and fix invalid tests
Cross-module struct initializers which set properties directly are deprecated in Swift 4.1, and will be an error in Swift 5.
The fact that a custom initializer was needed to construct the invalid Generators used by the test is indicative of a design flaw in the tests. If creating an invalid Generator with the library's default API is not possible, then testing such impossible-to-create invalid Generators is unnecessary. It is enough to test that attempts to create these invalid generators will fail.