Weakly link CryptoKit when building with the iOS 13 SDK
CryptoKit must be weakly linked so that code compiled with an SDK than can import CryptoKit does not crash when run on an OS version that pre-dates CryptoKit.
The weak link build setting must be conditional on the SDK so that the project can be built by Xcode 10, which is unable to link CryptoKit.